requirements.txt
rdkit
git+https://github.com/bp-kelley/descriptastorus
DeepPurpose
pytdc
tabulate
matplotlib
dgl
